hiberman: upgrade to syslog v6
We're trying to migrate from syslog v4, since it has a dependency
affected by RUSTSEC-2020-0071.

BUG=b:239449434
TEST=CQ

Cq-Depend: chromium:3781683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3781447
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>

+++ hiberman/src/hiberlog.rs
@@ -122,7 +122,7 @@ impl Hiberlog {
             flush_threshold: FLUSH_THRESHOLD,
             to_kmsg: false,
             out: HiberlogOut::Syslog,
-            pid: std::process::id(),
+            pid: std::process::id() as u32,
             syslogger,
         })
     }
@@ -485,7 +485,7 @@ fn create_syslogger() -> BasicLogger {
         facility: Facility::LOG_USER,
         hostname: None,
         process: "hiberman".into(),
-        pid: std::process::id() as i32,
+        pid: std::process::id() as u32,
     };
 
     let logger = syslog::unix(formatter).expect("Could not connect to syslog");
